Little more spiking
